+++ spec/lib/publisher_spec.rb
@@ -16,7 +16,7 @@ describe Publisher do
   describe '#text' do
     it 'is a formatted version of the prefill' do
       p = Publisher.new(alice, prefill: "@{alice; #{alice.diaspora_handle}}")
-      expect(p.text).to eq("alice")
+      expect(p.text).to eq("@alice")
     end
   end